Bug fix to command setup and command loop switching.
Added the missing call to initial command setup, which may have caused
initialization errors.
Fixed error for decision of which command is currently being executed,
so that it correctly chooses the current command.  Consequently, the
call to command loop was appropriately moved above that decision so as
to be always called for each command.