Bug fixes to Search code to remove compilation errors.
Fixed various misspelled names, missing function calls, and the like so
as to make sure the compilation is clean.
Modified break statements within the case statement so as to maintain
proper logic flow, since none of the cases were meant to drop through.
Moved Flame summary debugging to be grouped with the other summaries.